create section-banner component for section-news, community
